refactor(speech): extract supertonic tests to sibling file
Move the inline `#[cfg(test)] mod tests` out of supertonic.rs into a
sibling `supertonic/tests.rs`, mirroring the test-module extraction
series (#268–#293). supertonic.rs 554 → 402 (−152); new
supertonic/tests.rs is 151 lines.

The module is a non-root file, so the submodule resolves to
`supertonic/tests.rs` (mkdir'd). The test body has no depth-sensitive
relative paths (no include_str!/`"../`), so the cut is verbatim — only
the `mod tests {` wrapper is replaced with `mod tests;`. `use super::*`
reaches all parent privates unchanged.

Guarded under the matching feature build (a default `cargo test` does
not compile this module): `cargo test -p primer-speech --features
supertonic --lib` shows the identical count pre/post (43 passed, 1
ignored; 11 supertonic tests). Workspace 0 failures (51 ok blocks);
clippy + fmt clean. No behaviour change, test-only relocation.
